+++ src/routes/Check.tsx
@@ -1,4 +1,4 @@
-import { FinishKakaoLogin } from "../socialLogin";
+import { FinishKakaoLogin } from "../components/KakaoLogin";
 
 const Check = () => {
   const code = new URL(window.location.href).searchParams.get("code");
